requirements.txt
streamlit
streamlit-plotly-events

